requirements.txt
openai~=1.3.8
fastapi~=0.104.1
pydantic~=2.5.2
python-dotenv~=1.0.0
langchain
pypdf
tiktoken
requests
streamlit
wandb
uvicorn